Allow connect to reader after a reader reboot
Changes to allow connect to the reader after reboot. Differences between a short network fail and a reader reboot or long time without connection.
*Short network error: reconnects and ask for events.
*Long time without connection: Reconnects, configure and assign events

+++ SpeedwayReaderManager.cs
@@ -5,7 +5,7 @@ using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
-namespace WS.Reader.Procedure
+namespace Rhesus.WS.Procedure
 {
     internal class SpeedwayReaderManager
     {
@@ -13,9 +13,11 @@ namespace WS.Reader.Procedure
 
         private const string READER_USERNAME = "root";
         private const string READER_PASSWORD = "impinj";
+        private readonly TimeSpan MAX_DISCONNECT_TIME = new TimeSpan(0, 0, 35);
 
         //Properties
         private Reader ReaderInformation;
+        private ReaderParameters _activeReaderConfig;
         private readonly ushort ReaderConnectTimeoutSecs;
         private readonly uint KeepAliveSecs;
         private readonly ushort MaxKeepAliveLost;
@@ -24,11 +26,24 @@ namespace WS.Reader.Procedure
         //ImpinjAPI
         private ImpinjReader _impinjReader;
 
-        //Estado
-        internal bool Running { get; set; }
+        /**********ESTADO***********/
+        /// <summary>
+        /// Indicates when the reader is enabled and have to be running
+        /// </summary>
+        internal bool ReaderEnabled { get; set; }
+
+        /// <summary>
+        /// Indicates when the reader is configured
+        /// </summary>
+        private bool ReaderConfigured { get; set; }
+        
+        /// <summary>
+        /// Indicates when the reader is connected
+        /// </summary>
         internal bool IsConnected { get; set; }
-        private System.Timers.Timer _checkNetworkTimer;
         private int ReconnectRetries = 0;
+        private DateTime _disconnectTime = DateTime.Now.AddDays(-1);
+        private System.Timers.Timer _checkNetworkTimer;
 
         //Inventario
         internal bool InventoryRunning { get; private set; }
@@ -36,7 +51,7 @@ namespace WS.Reader.Procedure
         private static BlockingCollection<string> _tagInventoryThreadSafe;
 
         //GPIO
-        private Dictionary<UInt16, Boolean> GPIOPortStatus = new Dictionary<UInt16, Boolean>();
+        private Dictionary<ushort, bool> GPIOPortStatus = new Dictionary<ushort, bool>();
         internal bool GPIOPortStatus_1
         {
             get { return this.GPIOPortStatus != null && this.GPIOPortStatus.ContainsKey(1) && this.GPIOPortStatus[1]; }
@@ -46,15 +61,19 @@ namespace WS.Reader.Procedure
             get { return this.GPIOPortStatus != null && this.GPIOPortStatus.ContainsKey(2) && this.GPIOPortStatus[2]; }
         }
 
+        //Antenna Status
+        private Dictionary<ushort, bool> AntennaStatus = new Dictionary<ushort, bool>();
+
         //Event Handlers
         internal event EventHandler<ReaderEventArgs> GPIEvent;
         internal event EventHandler<ReaderEventArgs> TagReportEvent;
-        internal event EventHandler NewTagReadedEvent;
+        internal event EventHandler<ReaderEventArgs> NewTagReadedEvent;
         internal event EventHandler<ReaderEventArgs> InventoryFinishEvent;
 
-        public SpeedwayReaderManager(Reader readerInfo, ushort ReaderConnectTimeoutSecs, uint KeepAliveSecs, ushort MaxKeepAliveLost, ushort RebootAfterReconnects)
+        public SpeedwayReaderManager(Reader ReaderInformation, ushort ReaderConnectTimeoutSecs, uint KeepAliveSecs, ushort MaxKeepAliveLost, ushort RebootAfterReconnects)
         {
-            this.ReaderInformation = readerInfo;
+            this.ReaderInformation = ReaderInformation;
+            this._activeReaderConfig = this.ReaderInformation.T_READER_PARAMETERS[0];
             this.ReaderConnectTimeoutSecs = ReaderConnectTimeoutSecs;
             this.KeepAliveSecs = KeepAliveSecs;
             this.MaxKeepAliveLost = MaxKeepAliveLost;
@@ -65,11 +84,14 @@ namespace WS.Reader.Procedure
 
         #region Reader Connect
 
+        /// <summary>
+        /// Connect and Init reader
+        /// </summary>
         internal void StartReader()
         {
             try
             {
-                this.Running = true;
+                this.ReaderEnabled = true;
                 this.ReaderConnect();
                 this.InitReader();
             }
@@ -86,7 +108,7 @@ namespace WS.Reader.Procedure
         {
             try
             {
-                this.Running = false;
+                this.ReaderEnabled = false;
                 this.UnsetReaderEvents();
                 this.ReaderDisconnect();
             }
@@ -111,8 +133,7 @@ namespace WS.Reader.Procedure
                     this.IsConnected = true;
                     this.ReconnectRetries = 0;
                     logger.Info("[READER CONECTADO!!]: {0}-{1}", this._impinjReader.Name, this._impinjReader.Address);
-                    this.CheckReaderStatus();
-                    
+                    this.CheckStatusAfterReconnect();
                 }
                 catch (Exception ex)
                 {
@@ -126,38 +147,6 @@ namespace WS.Reader.Procedure
         }
 
         /// <summary>
-        /// Check reader status. Call this method after connect
-        /// </summary>
-        private void CheckReaderStatus()
-        {
-            Impinj.OctaneSdk.Status _readerStatus = this._impinjReader.QueryStatus();
-            //Check if reader is in inventory and stops it
-            if (!this.InventoryRunning && _readerStatus.IsSingulating)
-            {
-                logger.Warn("[INVENTORY STOPED WHILE READER WAS OFFLINE] Trying to stop");
-                this.StopInventory();
-            }
-
-            //Resume reader events
-            this._impinjReader.ResumeEventsAndReports();
-
-            //Log Reader temperature
-            logger.Info("READER TEMPERATURE: {0}ºC", _readerStatus.TemperatureInCelsius);
-
-            //Log Antenna information
-            foreach(Impinj.OctaneSdk.AntennaStatus _antStatus in _readerStatus.Antennas)
-            {
-                logger.Info("ANTENNA STATUS [PORT: {0} | CONNECTED: {1}]", _antStatus.PortNumber, _antStatus.IsConnected);
-            }
-
-            //Log GPI information
-            foreach(Impinj.OctaneSdk.GpiStatus _gpiStatus in _readerStatus.Gpis)
-            {
-                logger.Info("GPI STATUS [PORT: {0} | STATUS: {1}]", _gpiStatus.PortNumber, _gpiStatus.State);
-            }
-        }
-
-        /// <summary>
         /// Initialize Reader. Assign Events and configure parameters
         /// </summary>
         private void InitReader()
@@ -166,8 +155,11 @@ namespace WS.Reader.Procedure
             {
                 try
                 {
+                    //Delete AllOpSequences on first connect
+                    this._impinjReader.DeleteAllOpSequences();
                     this.InitializeReaderParameters();
                     this.SetReaderEvents();
+                    this.ReaderConfigured = true;
                 }
                 catch (Exception ex)
                 {
@@ -187,7 +179,7 @@ namespace WS.Reader.Procedure
             this.IsConnected = false;
             try
             {
-                this.StopCheckNetworkTimer();
+                this._disconnectTime = DateTime.Now;
                 _impinjReader.Disconnect();
             }
             catch(Exception ex)
@@ -201,21 +193,72 @@ namespace WS.Reader.Procedure
         {
             try
             {
-                if (!this.IsConnected && this.Running)
+                if (this.ReaderEnabled)
                 {
-                    if (this.Running)
+
+                    DateTime _reconnectTime = DateTime.Now;
+                    this.ReaderConfigured = _reconnectTime - this._disconnectTime < MAX_DISCONNECT_TIME;
+                    logger.Info("[READER CONFIGURED: {0}]", this.ReaderConfigured);
+                    if (!this.ReaderConfigured)
                     {
-                        this.ReaderConnect();
+                        this.StartReader();
                     }
                     else
                     {
-                        this.StartReader();
+                        this.ReaderConnect();
                     }
                 }
             }
             catch (Exception ex)
             {
                 logger.Error(ex, "[ERROR RECONNECT READER]");
+            }
+        }
+
+        /// <summary>
+        /// Check reader status. Call this method after connect
+        /// </summary>
+        private void CheckStatusAfterReconnect()
+        {
+            Impinj.OctaneSdk.Status _readerStatus = this._impinjReader.QueryStatus();
+            //Check if reader is in inventory and stops it
+            if (!this.InventoryRunning && _readerStatus.IsSingulating)
+            {
+                logger.Warn("[INVENTORY STOPED WHILE READER WAS OFFLINE] Trying to stop");
+                this.StopInventory();
+            }
+
+            //Resume reader events
+            this._impinjReader.ResumeEventsAndReports();
+
+            this.GetReaderStatus(_readerStatus);
+        }
+
+        /// <summary>
+        /// Log Reader state and save antena and GPI state
+        /// </summary>
+        /// <param name="_readerStatus"></param>
+        private void GetReaderStatus(Impinj.OctaneSdk.Status _readerStatus = null)
+        {
+            if(_readerStatus == null)
+            {
+                _readerStatus = this._impinjReader.QueryStatus();
+            }
+            //Log Reader temperature
+            logger.Info("READER TEMPERATURE: {0}ºC", _readerStatus.TemperatureInCelsius);
+
+            //Log Antenna information
+            foreach (Impinj.OctaneSdk.AntennaStatus _antStatus in _readerStatus.Antennas)
+            {
+                this.AntennaStatus[_antStatus.PortNumber] = _antStatus.IsConnected;
+                logger.Info("ANTENNA STATUS [PORT: {0} | CONNECTED: {1}]", _antStatus.PortNumber, _antStatus.IsConnected);
+            }
+
+            //Log GPI information
+            foreach (Impinj.OctaneSdk.GpiStatus _gpiStatus in _readerStatus.Gpis)
+            {
+                this.SavePortSatus(_gpiStatus.PortNumber, _gpiStatus.State);
+                logger.Info("GPI STATUS [PORT: {0} | STATUS: {1}]", _gpiStatus.PortNumber, _gpiStatus.State);
             }
         }
 
@@ -226,17 +269,29 @@ namespace WS.Reader.Procedure
         private void InitializeReaderParameters()
         {
             logger.Info("[CONFIGURANDO READER]");
-            _impinjReader.DeleteAllOpSequences();
-
-            // Get the default settings
-            // We'll use these as a starting point
-            // and then modify the settings we're interested in.
+            
             Settings settings = _impinjReader.QueryDefaultSettings();
-            var parameters = ReaderInformation.T_READER_PARAMETERS[0];
 
-            // Set the start trigger to Immediate.
-            // This will allow the reader to start as soon as it boots up.
-            //settings.AutoStart.Mode = AutoStartMode.Immediate;
+            this.ConfigReader_InventoryConfig(this._activeReaderConfig, settings);
+            this.ConfigReader_Antennas(this._activeReaderConfig, settings);
+            this.ConfigReader_Keepalive(settings);
+            
+
+            // If this application disconnects from the 
+            // reader, hold all tag reports and events.
+            settings.HoldReportsOnDisconnect = true;
+            this._impinjReader.ApplySettings(settings);
+        }
+
+        private void ConfigReader_InventoryConfig(ReaderParameters ReaderParameters, Impinj.OctaneSdk.Settings settings = null)
+        {
+            if(settings == null)
+            {
+                settings = this._impinjReader.QuerySettings();
+            }
+
+            //No Autostop
+            settings.AutoStop.Mode = AutoStopMode.None;
 
             // Tell the reader to include some paremeters
             // in all tag reports.
@@ -246,25 +301,45 @@ namespace WS.Reader.Procedure
             settings.Report.IncludeSeenCount = true;
             settings.Report.IncludePcBits = true;
 
-            settings.Antennas.EnableAll();
-
             // Send a tag report for every tag read.
             settings.Report.Mode = ReportMode.Individual;
 
             settings.ReaderMode = ReaderMode.AutoSetDenseReader;
             settings.SearchMode = SearchMode.DualTarget;
-            settings.Session = (ushort)parameters.SESSION_;
+            settings.Session = (ushort)ReaderParameters.SESSION_;
 
-            settings.Antennas.EnableAll();
-            for (ushort i = 1; i <= settings.Antennas.Length; i++)
+            //Apply configuration
+            this._impinjReader.ApplySettings(settings);
+        }
+
+        private void ConfigReader_Antennas(ReaderParameters ReaderParameters, Impinj.OctaneSdk.Settings settings = null)
+        {
+            if (settings == null)
             {
-                if (settings.Antennas.GetAntenna(i).IsEnabled)
+                settings = this._impinjReader.QuerySettings();
+            }
+
+            foreach (Impinj.OctaneSdk.AntennaConfig _antCfg in settings.Antennas)
+            {
+                bool IsAntennaConnected;
+                bool _getValue = this.AntennaStatus.TryGetValue(_antCfg.PortNumber, out IsAntennaConnected);
+                _antCfg.IsEnabled = !_getValue || IsAntennaConnected;
+                if (_antCfg.IsEnabled)
                 {
-                    // Set the Transmit Power and Receive Sensitivity
-                    // settings.Antennas.GetAntenna(i).RxSensitivityInDbm = parameters.noise;
-                    settings.Antennas.GetAntenna(i).TxPowerInDbm = (double)parameters.RFLEVEL / 100;
+                    _antCfg.TxPowerInDbm = (double)ReaderParameters.RFLEVEL / 100;
                 }
             }
+            //Apply configuration
+            this._impinjReader.ApplySettings(settings);
+        }
+
+        private void ConfigReader_Keepalive(Impinj.OctaneSdk.Settings settings = null)
+        {
+            if (settings == null)
+            {
+                settings = this._impinjReader.QuerySettings();
+            }
+
             //KeepAlive configuration
             KeepaliveConfig _keepaliveCfg = new KeepaliveConfig();
             _keepaliveCfg.Enabled = true;
@@ -273,32 +348,26 @@ namespace WS.Reader.Procedure
             _keepaliveCfg.LinkDownThreshold = this.MaxKeepAliveLost;
             settings.Keepalives = _keepaliveCfg;
 
-            // If this application disconnects from the 
-            // reader, hold all tag reports and events.
-            settings.HoldReportsOnDisconnect = true;
-
-            //No Autostop
-            settings.AutoStop.Mode = AutoStopMode.None;
-
-            // Apply the newly modified settings.
-            _impinjReader.ApplySettings(settings);
+            //Apply configuration
+            this._impinjReader.ApplySettings(settings);
         }
 
         private void SetReaderEvents()
         {
+            this.UnsetReaderEvents();
             logger.Info("[ASIGNACION EVENTOS READER]");
-            //Eventos de conexion
             this._impinjReader.ConnectionLost += _impinjReader_ConnectionLost;
             this._impinjReader.KeepaliveReceived += _impinjReader_KeepaliveReceived;
-
-            //Eventos de GPI
+            this._impinjReader.AntennaChanged += _impinjReader_AntennaChanged;
             this._impinjReader.GpiChanged += OnGPIChanged;
         }
 
         private void UnsetReaderEvents()
         {
+            logger.Info("[ELIMINAR ASIGNACION EVENTOS READER]");
             this._impinjReader.ConnectionLost -= _impinjReader_ConnectionLost;
             this._impinjReader.KeepaliveReceived -= _impinjReader_KeepaliveReceived;
+            this._impinjReader.AntennaChanged -= _impinjReader_AntennaChanged;
             this._impinjReader.GpiChanged -= OnGPIChanged;
         }
 
@@ -363,7 +432,7 @@ namespace WS.Reader.Procedure
                     {
                         logger.Info("[UNIQUE TAG FOUND]: {0}", _tagEPC);
                         readerManager.ReadedTags.Add(_tagEPC);
-                        readerManager.NewTagReadedEvent?.Invoke(readerManager, new EventArgs());
+                        readerManager.NewTagReadedEvent?.Invoke(readerManager, new ReaderEventArgs { FoundEpc = _tagEPC });
                     }
                 }
                 _tagInventoryThreadSafe.Dispose();
@@ -395,8 +464,16 @@ namespace WS.Reader.Procedure
             //logger.Trace("[KeepAliveReceived]: {0}-{1}", reader.Name, reader.Address);
         }
 
+        private void _impinjReader_AntennaChanged(ImpinjReader reader, AntennaEvent e)
+        {
+            logger.Info("[ANTENNA CHANGED EVENT] - ANTENNA: {0} | STATE: {1}", e.PortNumber, e.State);
+            this.GetReaderStatus();
+            this.ConfigReader_Antennas(this._activeReaderConfig);
+        }
+
         private void OnGPIChanged(ImpinjReader reader, GpiEvent e)
         {
+            logger.Info("GPIO Event[PUERTO: {0} ESTADO: {1}]", e.PortNumber, e.State);
             this.SavePortSatus(e.PortNumber, e.State);
             ReaderEventArgs _evtArgs = new ReaderEventArgs { GPIPortNumber = e.PortNumber, GPIPortStatus = e.State };
             GPIEvent?.Invoke(this, _evtArgs);
@@ -442,7 +519,6 @@ namespace WS.Reader.Procedure
                 this.GPIOPortStatus = new Dictionary<ushort, bool>();
             }
             this.GPIOPortStatus[PortNumber] = state;
-            logger.Info("GPIO Event[PUERTO: {0} ESTADO: {1}]", PortNumber, state);
         }
 
         #region Network Timer
@@ -458,26 +534,36 @@ namespace WS.Reader.Procedure
 
         private void _checkNetworkTimer_Elapsed(object sender, System.Timers.ElapsedEventArgs e)
         {
-            bool ChkNetwork = NetworkUtils.DoPing(this.ReaderInformation.IP);
-            if (ChkNetwork)
+            if (this.ReaderEnabled)
+            {
+                bool ChkNetwork = NetworkUtils.DoPing(this.ReaderInformation.IP);
+                if (ChkNetwork)
+                {
+                    this.StopCheckNetworkTimer();
+
+                    //Reboot if reconnects exceed
+                    if (this.RebootAfterReconnects > 0 && this.ReconnectRetries > this.RebootAfterReconnects)
+                    {
+                        this.ReconnectRetries = 0;
+                        this.SendRShellCommand("reboot");
+                    }
+                    this.ReconnectAfterConnectionLost();
+                }
+            }
+            else
             {
                 this.StopCheckNetworkTimer();
-
-                //Reboot if reconnects exceed
-                if(this.RebootAfterReconnects > 0 && this.ReconnectRetries > this.RebootAfterReconnects)
-                {
-                    this.ReconnectRetries = 0;
-                    this.SendRShellCommand("reboot");
-                }
-                this.ReconnectAfterConnectionLost();
             }
         }
 
         private void StopCheckNetworkTimer()
         {
-            this._checkNetworkTimer.AutoReset = false;
-            this._checkNetworkTimer.Enabled = false;
-            this._checkNetworkTimer.Dispose();
+            if (this._checkNetworkTimer != null)
+            {
+                this._checkNetworkTimer.AutoReset = false;
+                this._checkNetworkTimer.Enabled = false;
+                this._checkNetworkTimer.Dispose();
+            }
         }
         #endregion
